// src/modules/crud-module/crud-table/data/schema.ts
import { z } from "zod"

// We're keeping a simple non-relational schema here.
// IRL, you will have a schema for your data models.
export const taskSchema = z.object({
  userId: z.string(),
  name: z.string(),
  role: z.string(),
  phoneNumber: z.string(),
  email: z.string(),
  status: z.string(),
  label: z.string(),
})

export type Task = z.infer<typeof taskSchema>